@@ -21,7 +21,7 @@ class PortfoliosController < ApplicationController
 
     @user = User.find(current_user.id)
     @portfolios = @user.portfolios
-    
+
   end
 
   # GET /portfolios/1
@@ -47,19 +47,23 @@ class PortfoliosController < ApplicationController
   # POST /portfolios
   # POST /portfolios.json
   def create
-    @portfolio = Portfolio.new(portfolio_params)
-    # @stock = Stock.find_by(symbol: params[:portfolio][:holdings_attributes][:symbol])
-    # @holding = Holding.find()
 
-    respond_to do |format|
-      if @portfolio.save
-        format.html { redirect_to @portfolio, notice: 'Portfolio was successfully created.' }
-        format.json { render :show, status: :created, location: @portfolio }
-      else
-        format.html { render :new }
-        format.json { render json: @portfolio.errors, status: :unprocessable_entity }
-      end
-    end
+    @portfolio = Portfolio.create(portfolio_params)
+    @user = User.find(current_user.id)
+    @user.portfolios << @portfolio
+    StocksUser.delete_all
+    redirect_to portfolios_path
+
+
+    # respond_to do |format|
+    #   if @portfolio.save
+    #     format.html { redirect_to @portfolio, notice: 'Portfolio was successfully created.' }
+    #     format.json { render :show, status: :created, location: @portfolio }
+    #   else
+    #     format.html { render :new }
+    #     format.json { render json: @portfolio.errors, status: :unprocessable_entity }
+    #   end
+    # end
   end
 
   # PATCH/PUT /portfolios/1
