@@ -1,10 +1,11 @@
 class StocksUsersController < ApplicationController
   before_action :set_stock_user, only: [:show, :edit, :update, :destroy]
+  skip_before_filter :verify_authenticity_token
 
   # GET /stock_users
   # GET /stock_users.json
   def index
-    @stock_users = StocksUser.all
+    @stocks_users = StocksUser.all
   end
 
   # GET /stock_users/1
@@ -15,7 +16,7 @@ class StocksUsersController < ApplicationController
 
   # GET /stock_users/new
   def new
-    @stock_user = StocksUser.new
+    @stocks_user = StocksUser.new
   end
 
   # GET /stock_users/1/edit
@@ -25,17 +26,29 @@ class StocksUsersController < ApplicationController
   # POST /stock_users
   # POST /stock_users.json
   def create
-    @stock_user = StocksUser.new(stock_user_params)
 
-    respond_to do |format|
-      if @stock_user.save
-        format.html { redirect_to @stock_user, notice: 'Stock user was successfully created.' }
-        format.json { render :show, status: :created, location: @stock_user }
-      else
-        format.html { render :new }
-        format.json { render json: @stock_user.errors, status: :unprocessable_entity }
-      end
+    if request.xhr?
+       @stocks_user = StocksUser.new(stocks_user_params)
+       if @stocks_user.save
+         render json: "200"
+       else
+          render json: "500"
+       end
+    else
+      @stocks_user = StocksUser.new(stocks_user_params)
+      redirect_to @stock_user
     end
+
+
+    # respond_to do |format|
+    #   if @stock_user.save
+    #     format.html { redirect_to @stock_user, notice: 'Stock user was successfully created.' }
+    #     format.json { render :show, status: :created, location: @stock_user }
+    #   else
+    #     format.html { render :new }
+    #     format.json { render json: @stock_user.errors, status: :unprocessable_entity }
+    #   end
+    # end
   end
 
   # PATCH/PUT /stock_users/1
@@ -65,11 +78,11 @@ class StocksUsersController < ApplicationController
   private
     # Use callbacks to share common setup or constraints between actions.
     def set_stock_user
-      @stock_user = StocksUser.find(params[:id])
+      @stocks_user = StocksUser.find(params[:id])
     end
 
     # Never trust parameters from the scary internet, only allow the white list through.
-    def stock_user_params
+    def stocks_user_params
       params.require(:stock_user).permit(:user_id, :stock_id)
     end
 end
