@@ -1,3 +1,4 @@
 # Be sure to restart your server when you modify this file.
 
-Rails.application.config.session_store :cookie_store, key: '_pocketboard_session'
+Rails.application.config.session_store :active_record_store
+
