@@ -2,7 +2,7 @@ class CreatePortfolios < ActiveRecord::Migration
   def change
     create_table :portfolios do |t|
       t.string :name
-      t.references :user, index: true, foreign_key: true
+      t.integer :user_id
 
       t.timestamps null: false
     end
